+++ index.js
@@ -27,8 +27,10 @@ export default class ServerlessPluginMonorepoNPMWorkspaces {
 
   beforePackage() {
     this.utils.log(PLUGIN_NAME, 'starting application packaging');
-
-    const { workspaceRootDirectoryPath = DEFAULT_WORKSPACE_ROOT_DIRECTORY_PATH } = this.serverless.service.initialServerlessConfig[PLUGIN_NAME];
+    if (!this.serverless.version.startsWith('3')) {
+      this.utils.log(`This plugin has only been tested with serverless v3.x, it may not be working for your version ${this.serverless.version}`);
+    }
+    const workspaceRootDirectoryPath = this.serverless.service.initialServerlessConfig[PLUGIN_NAME]?.workspaceRootDirectoryPath || DEFAULT_WORKSPACE_ROOT_DIRECTORY_PATH;
     const workspacePackageJsonPath = `${workspaceRootDirectoryPath}/package.json`;
     if (!fs.existsSync(workspacePackageJsonPath)) {
       throw new this.serverless.classes.Error(`Unable to find package.json at workspaceRootDirectoryPath ${workspaceRootDirectoryPath}`);
@@ -47,10 +49,12 @@ export default class ServerlessPluginMonorepoNPMWorkspaces {
     fs.rmSync(`${workspaceNodeModulesPath}_tmp`, { recursive: true, force: true });
     fs.renameSync(workspaceNodeModulesPath, `${workspaceNodeModulesPath}_tmp`);
 
+    // read package.json
+    const currentPackageJson = JSON.parse(fs.readFileSync('./package.json'));
+
     // execute command npm ci --workspace=my_package
-    const appName = process.cwd().split('/').reverse()[0];
-    this.utils.log(PLUGIN_NAME, `generating application ${appName} dependencies`);
-    spawnSync(npmCommand, ['ci', `--workspace=${process.cwd().split('/').reverse()[0]}`]);
+    this.utils.log(PLUGIN_NAME, `generating application ${currentPackageJson.name} dependencies`);
+    spawnSync(npmCommand, ['ci', `--workspace=${currentPackageJson.name}`]);
 
     // get symlinks from node_modules
     const workspaceNodeModules = fs.readdirSync(workspaceNodeModulesPath, {
@@ -58,29 +62,100 @@ export default class ServerlessPluginMonorepoNPMWorkspaces {
     });
     const symlinks = [];
     workspaceNodeModules.forEach((entry) => {
-      if (entry.isSymbolicLink()) {
-        symlinks.push(entry);
+      // scoped packages
+      if (entry.name.startsWith('@')) {
+        const scopePackage = fs.readdirSync(`${entry.parentPath}/${entry.name}`, {
+          withFileTypes: true,
+        });
+        scopePackage.forEach((subPackage) => {
+          if (subPackage.isSymbolicLink()) {
+            symlinks.push(subPackage);
+            symlinks.push({
+              ...subPackage,
+              fullName: `${entry.name}/${subPackage.name}`,
+            });
+          }
+        });
+      } else if (entry.isSymbolicLink()) {
+        symlinks.push({
+          ...entry,
+          fullName: entry.name,
+        });
       }
     });
 
-    const { layerPath = DEFAULT_LAYER_PATH } = this.serverless.service.initialServerlessConfig[PLUGIN_NAME];
+    const layerPath = this.serverless.service.initialServerlessConfig[PLUGIN_NAME]?.layerPath || DEFAULT_LAYER_PATH;
     // copy node_modules from root dir without symlinks to layers/main/nodejs/node_modules
     fs.rmSync(layerPath, { recursive: true, force: true });
     fs.cpSync(workspaceNodeModulesPath, `${layerPath}/nodejs/node_modules`, {
       recursive: true,
       filter: (source) => !symlinks.map((entry) => `${entry.parentPath}/${entry.name}`).includes(source),
     });
-    // copy internal deps that where symlinks in root dir to node_modules
-    // read package.json deps and copy those that where symlinks in root node_modules to current node_modules
-    const currentPackageJson = JSON.parse(fs.readFileSync('./package.json'));
-    Object.keys(currentPackageJson.dependencies).forEach((dep) => {
-      if (symlinks.find((link) => link.name === dep)) {
-        fs.cpSync(`../${dep}`, `./node_modules/${dep}`, {
-          recursive: true,
-          force: true,
-        });
+
+    // get a map of directories in workspaces with dirname <> package name
+    const packageNames = new Map();
+    const workspacePackages = fs.readdirSync('..', {
+      withFileTypes: true,
+    });
+
+    workspacePackages.forEach((workspacePackage) => {
+      const packageJson = `../${workspacePackage.name}/package.json`;
+      if (workspacePackage.isDirectory() && fs.existsSync(packageJson)) {
+        const packageJsonDir = JSON.parse(fs.readFileSync(packageJson));
+        packageNames.set(packageJsonDir.name, { packageJson: packageJsonDir, workspacePackageName: workspacePackage.name });
       }
     });
+
+    // copy internal deps that where symlinks in root dir to node_modules
+    const toIgnore = ['.gitignore'];
+    if (this.serverless.service.initialServerlessConfig.package?.patterns?.length) {
+      // eslint-disable-next-line no-unsafe-optional-chaining
+      toIgnore.push(...this.serverless.service.initialServerlessConfig.package?.patterns.map((pattern) => {
+        let sanitizedPattern = pattern.startsWith('!') ? pattern.substring(1) : pattern;
+        if (sanitizedPattern.endsWith('/**')) {
+          sanitizedPattern = sanitizedPattern.substring(0, sanitizedPattern.length - 3);
+        }
+        return sanitizedPattern;
+      }));
+    }
+    if (fs.existsSync(`${workspaceRootDirectoryPath}/.gitignore`)) {
+      const workspaceGitIgnore = fs.readFileSync(`${workspaceRootDirectoryPath}/.gitignore`, { encoding: 'utf8' });
+      toIgnore.push(...workspaceGitIgnore.split('\n'));
+    }
+    if (fs.existsSync('.gitignore')) {
+      const currentGitIgnore = fs.readFileSync('.gitignore', { encoding: 'utf8' });
+      toIgnore.push(...currentGitIgnore.split('\n'));
+    }
+
+    function copyMonorepoDependencies(packageJson) {
+      Object.keys(packageJson.dependencies).forEach((dep) => {
+        if (symlinks.find((link) => link.fullName === dep)) {
+          const packageName = `../${packageNames.get(dep).workspacePackageName}`;
+          const depToIgnore = [];
+          if (fs.existsSync(`${packageName}/.gitignore`)) {
+            depToIgnore.push(fs.readFileSync(`${packageName}/.gitignore`, { encoding: 'utf8' }).split('\n'));
+          }
+          fs.cpSync(`../${packageNames.get(dep).workspacePackageName}`, `./node_modules/${dep}`, {
+            recursive: true,
+            force: true,
+            filter: (source) => {
+              const filePath = source.split('/');
+              const toIgnoreAll = [...toIgnore, ...depToIgnore];
+              let res = true;
+              toIgnoreAll.forEach((tia) => {
+                if (filePath.some((t) => t === tia)) {
+                  res = false;
+                }
+              });
+              return res;
+            },
+          });
+          copyMonorepoDependencies(packageNames.get(dep).packageJson);
+        }
+      });
+    }
+
+    copyMonorepoDependencies(currentPackageJson);
 
     // remove node_modules generated in root dir
     this.utils.log(PLUGIN_NAME, 'removing temporary files');