+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release history
 
+## 1.1.0
+- fix error undefined if no serverless-plugin-monorepo-npm-workspaces default conf in serverless.yml
+- fix error if package directory name in monorepo is different from package.json name property when copying dependencies
+- properly handle scoped packages when manipulating symlinks
+- use .gitignore and serverless.yml package.patterns conf to skip copying unwanted files
+
 ## 1.0.1
 - updated readme
 - added license